fixup! core: rework IP configuration in NetworkManager using layer 3 configuration
The NML3Cfg instance is our manager object for the IP configuration of
an ifindex (netdev device). There is supposed to be only one, which is
ensured via nm_netns_access_l3cfg().

Likewise, when we expose the IP configuration on D-Bus, it makes no
sense to have multiple NMIPConfig instances. Instead, let any user also
share a unique instance, via nm_l3cfg_ipconfig_acquire().

Note that this changes the meaning of what NMIPXConfig previously meant
for VPN connections. It's not actually clear what it meant, it kinda
exported some of the configurations. But that makes no sense to me. We
should expose the per-ifindex IP info, and all that is needed for that
is accessible via NML3Cfg.